% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blbglm}
\alias{blbglm}
\title{BLBGLM}
\usage{
blbglm(formula, data, m = 10, B = 5000, family)
}
\arguments{
\item{formula}{regression formula to fit}

\item{data}{data}

\item{m}{splitted data size}

\item{B}{times of bootstrap}

\item{family}{family in glm}
}
\description{
BLBGLM
}
