% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm-package}
\title{blblm: Bag of Little Bootstrap for Regression Models}
\description{
Use the bag of little bootstrap algorithm to fit linear regression model and GLM and return coeficeients and sigmas of the model.
}
\details{
Regression models with Little Bag of Bootstraps
}
