% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{BLB Linear regression}
\usage{
blblm(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{regression formula to fit}

\item{data}{data or data folder directory}

\item{m}{splitted data size}

\item{B}{times of bootstrap}
}
\description{
BLB Linear regression
}
