% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{confidence interval for each terms}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{fitted model}

\item{parm}{TRUE/FALSE}

\item{level}{confidence level}

\item{...}{extra conditions}
}
\description{
confidence interval for each terms
}
