% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{glm1}
\alias{glm1}
\title{estimate the GLM estimates based on given number of repetitions}
\usage{
glm1(formula, data, freqs, family)
}
\arguments{
\item{formula}{regression formula to fit}

\item{data}{data}

\item{freqs}{weights for fitting}

\item{family}{family in glm}
}
\description{
estimate the GLM estimates based on given number of repetitions
}
