% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{estimate the linear regression estimates based on given the number of repetitions}
\usage{
lm1(formula, data, freqs)
}
\arguments{
\item{formula}{regression formula to fit}

\item{data}{data}

\item{freqs}{weights for fitting}
}
\description{
estimate the linear regression estimates based on given the number of repetitions
}
