% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{compute the estimates for linear regression}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{regression formula to fit}

\item{data}{data}

\item{n}{number of random vectors to draw}

\item{B}{times of bootstrap}
}
\description{
compute the estimates for linear regression
}
