% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{par_blblm}
\alias{par_blblm}
\title{BLB Linear regression with paralization.}
\usage{
par_blblm(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{regression formula to fit}

\item{data}{data}

\item{m}{splitted data size}

\item{B}{times of bootstrap}
}
\description{
BLB Linear regression with paralization.
}
