% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{predict the model by fiven new data}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{fitted model}

\item{new_data}{data}

\item{confidence}{True/False}

\item{level}{confidence level}

\item{...}{extra conditions}
}
\description{
predict the model by fiven new data
}
