% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{sigma for blb regression model}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{fitted model}

\item{confidence}{True/False}

\item{level}{confidence level}

\item{...}{extra conditions}
}
\description{
sigma for blb regression model
}
